Fixed module loading issues
When using the module path for referring to functions (@triptease/html-jsx) the `raw` method references two different instances of the Raw class. One from the module and the other from sources. This causes instanceOf checks to fail and prevents raw from working correctly.

By using a relative import path this problem is solved.